+++ CHANGELOG.rst
@@ -31,6 +31,8 @@ Changelog
   :class:`~cryptography.x509.ocsp.OCSPResponse`. 
 * Restored support for signing certificates and other structures in
   :doc:`/x509/index` with SHA3 hash algorithms.
+* :class:`~cryptography.hazmat.primitives.ciphers.algorithms.TripleDES` is
+  disabled in FIPS mode.
 
 .. _v36-0-1: